requirements.txt
fastapi
uvicorn[standard]
openai-whisper
openai>=1.0
python-multipart
python-dotenv
torch
